backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pypdf
langchain
pinecone
sentence-transformers
transformers
langchain-core
langgraph
langchain-groq
python-dotenv
bs4









